divided the rocognition picture and the medicines' interactions into folders
