+++ PrescriptionDAL/MedicineInteraction/FullInteractionType.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PrescriptionDAL.MedicineInteraction
+{
+    class FullInteractionType
+    {
+        public string comment { get; set; }
+        public List<MinConcept> minConcept { get; set; }
+        public List<InteractionPair> interactionPair { get; set; }
+    }
+}
