+++ PrescriptionDAL/MedicineInteraction/InteractionConcept.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PrescriptionDAL.MedicineInteraction
+{
+    class InteractionConcept
+    {
+        public MinConceptItem minConceptItem { get; set; }
+        public SourceConceptItem sourceConceptItem { get; set; }
+    }
+}
