+++ PrescriptionDAL/MedicineInteraction/MinConceptItem.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PrescriptionDAL.MedicineInteraction
+{
+    class MinConceptItem
+    {
+        public string rxcui { get; set; }
+        public string name { get; set; }
+        public string tty { get; set; }
+    }
+}
