+++ PrescriptionDAL/MedicineInteraction/RootMedicionInteraction.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PrescriptionDAL.MedicineInteraction
+{
+    class RootMedicionInteraction
+    {
+        public string nlmDisclaimer { get; set; }
+        public UserInput userInput { get; set; }
+        public List<FullInteractionTypeGroup> fullInteractionTypeGroup { get; set; }
+    }
+}
