+++ PrescriptionDAL/MedicineInteraction/SourceConceptItem.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PrescriptionDAL.MedicineInteraction
+{
+    class SourceConceptItem
+    {
+        public string id { get; set; }
+        public string name { get; set; }
+        public string url { get; set; }
+    }
+}
