+++ PrescriptionDAL/MedicineInteraction/UserInput.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PrescriptionDAL.MedicineInteraction
+{
+    class UserInput
+    {
+        public List<string> sources { get; set; }
+        public List<string> rxcuis { get; set; }
+    }
+}
