+++ PrescriptionDAL/RegonitionPicture/Result.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PrescriptionDAL.RegonitionPicture
+{
+    class Result
+    {
+        public List<Tag> tags { get; set; }
+    }
+}
