+++ PrescriptionDAL/RegonitionPicture/Root.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PrescriptionDAL.RegonitionPicture
+{
+    class Root
+    {
+        public Result result { get; set; }
+        public Status status { get; set; }
+    }
+}
