+++ PrescriptionDAL/RegonitionPicture/Status.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PrescriptionDAL.RegonitionPicture
+{
+    class Status
+    {
+        public string text { get; set; }
+        public string type { get; set; }
+    }
+}
