+++ PrescriptionDAL/RegonitionPicture/Tag.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PrescriptionDAL.RegonitionPicture
+{
+    class Tag
+    {
+        public double confidence { get; set; }
+        public Tag2 tag { get; set; }
+    }
+}
