+++ PrescriptionDAL/RegonitionPicture/Tag2.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PrescriptionDAL.RegonitionPicture
+{
+    class Tag2
+    {
+        public string en { get; set; }
+    }
+}
